refactor: 파일 이동
- model에 있던 파일 -> controller로 이동
- SetLocalStorage
- SetProductAdd
- SetVendingMachineCharge
- SetVendingMachinePurchase
